refactoring in process
